+++ src/map/party.cpp
@@ -1039,7 +1039,7 @@ int party_send_logout(map_session_data *sd)
 	return 1;
 }
 
-int party_send_message(map_session_data *sd,const char *mes,int len)
+int party_send_message(map_session_data *sd,const char *mes, size_t len)
 {
 	if(sd->status.party_id == 0)
 		return 0;
